+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Fixed bug in `_calc_branch_power_max` where `c_rating_b` was being used
 - Add support for storage to OPF_OLTC
 - Add wye-connected CapControl for ACP, ACR, LinDist3Flow, FBS and FOT formulations